change some double to float

+++ MandelBrot3D/JB_MandelModelController.h
@@ -29,7 +29,7 @@ typedef struct
 	int Iter[801][601], iter;
 	RGBA Color[801][601];
 	double V[801][601];
-	double S[801][601];
+	float S[801][601];
 	double dx, dy;
 	
 	NSThread *CalcThread;
@@ -42,7 +42,7 @@ typedef struct
 }
 -(int)getNum:(int)x :(int)y;
 -(RGBA)getColor:(int)x :(int)y;
--(double)getH:(int)x :(int)y;
+-(float)getH:(int)x :(int)y;
 -(void)newRange:(NSPoint)O :(NSPoint)E;
 -(void)Reset;
 @end
